创建门店目标数据来源模型并添加维护页面;修改门店目标逻辑,创建两条数据供展示.

+++ e2yun_addons/odoo12/outbound_report/models/target_completion.py
@@ -15,20 +15,25 @@ class TargetCompletion(models.Model):
             for rec in final:
                 rec.unlink()
         res = super(TargetCompletion, self).create(vals_list)
-        new_val_list = []
-        if res.jiesuan_amount:
-            jiesuan_dict = vals_list.copy()
-            jiesuan_dict.update({'completion': res.jiesuan_amount, 'stage_id': 1})
-            new_val_list.append(jiesuan_dict)
-        if res.target_amount:
-            target_dict = vals_list.copy()
-            target_dict.update({'completion': res.target_amount, 'stage_id': 2})
-            new_val_list.append(target_dict)
-        aa = super(TargetCompletion, self).create(new_val_list)
-        if aa:
-            kk = aa[0]
-            res.unlink()
-            return kk
+        # 判断是查询目标完成占比报表,生成两条数据
+        if 'target_year' in vals_list:
+            new_val_list = []
+            if res.jiesuan_amount:
+                jiesuan_dict = vals_list.copy()
+                jiesuan_dict.update({'completion': res.jiesuan_amount, 'stage_id': 1})
+                new_val_list.append(jiesuan_dict)
+            if res.target_amount:
+                target_dict = vals_list.copy()
+                target_dict.update({'completion': res.target_amount, 'stage_id': 2})
+                new_val_list.append(target_dict)
+            aa = super(TargetCompletion, self).create(new_val_list)
+            if aa:
+                kk = aa[0]
+                res.unlink()
+                return kk
+        # 不是,直接返回
+        else:
+            return res
 
     def default_target_year(self):
         ctx = self._context.copy()
@@ -53,70 +58,78 @@ class TargetCompletion(models.Model):
     @api.depends('target_year', 'target_month', 'kunnr', 'ywy')
     def _compute_target_amount(self):
         for rec in self:
-            # 选择月份（展示门店/业务月度目标）
-            if rec.target_month:
-                if rec.ywy:
-                    target_detail = self.env['team.target.detail'].search([('current_team_id', '=', rec.kunnr.id),
-                                                                           ('detail_year', '=', rec.target_year),
-                                                                           ('target_month', '=?', rec.target_month),
-                                                                           ('sales_member.id', '=?', rec.ywy.id)])
-                else:
-                    target_detail = self.env['team.target.detail'].search([('current_team_id', '=', rec.kunnr.id),
-                                                                          ('detail_year', '=', rec.target_year),
-                                                                          ('target_month', '=?', rec.target_month)])
-                target_amount = 0
-                if target_detail:
-                    for detail in target_detail:
-                        target_amount += detail.team_target_monthly
-                rec.target_amount = target_amount
-            # 不选择月份（展示门店/业务年度目标）
-            else:
-                # 选择业务员,年度目标通过分配目标的取值
-                if rec.ywy:
-                    ywy = rec.ywy.id
-                    target_detail = self.env['team.target.detail'].search([('current_team_id', '=', rec.kunnr.id),
-                                                                           ('detail_year', '=', rec.target_year),
-                                                                           ('sales_member.id', '=?', ywy)])
+            # 确认是查询目标完成占比报表
+            if rec.target_year:
+                # 选择月份（展示门店/业务月度目标）
+                if rec.target_month:
+                    if rec.ywy:
+                        target_detail = self.env['team.target.detail'].search([('current_team_id', '=', rec.kunnr.id),
+                                                                               ('detail_year', '=', rec.target_year),
+                                                                               ('target_month', '=?', rec.target_month),
+                                                                               ('sales_member.id', '=?', rec.ywy.id)])
+                    else:
+                        target_detail = self.env['team.target.detail'].search([('current_team_id', '=', rec.kunnr.id),
+                                                                              ('detail_year', '=', rec.target_year),
+                                                                              ('target_month', '=?', rec.target_month)])
                     target_amount = 0
                     if target_detail:
                         for detail in target_detail:
                             target_amount += detail.team_target_monthly
                     rec.target_amount = target_amount
-                # 不选择业务员,直接取门店年度目标
+                # 不选择月份（展示门店/业务年度目标）
                 else:
-                    target_year = self.env['team.target.year'].search([('team_id', '=', rec.kunnr.id),
-                                                                       ('target_year', '=', rec.target_year)])
-                    target_amount = 0
-                    if target_year:
-                        for target in target_year:
-                            target_amount += target.invoiced_target_year
-                    rec.target_amount = target_amount
+                    # 选择业务员,年度目标通过分配目标的取值
+                    if rec.ywy:
+                        ywy = rec.ywy.id
+                        target_detail = self.env['team.target.detail'].search([('current_team_id', '=', rec.kunnr.id),
+                                                                               ('detail_year', '=', rec.target_year),
+                                                                               ('sales_member.id', '=?', ywy)])
+                        target_amount = 0
+                        if target_detail:
+                            for detail in target_detail:
+                                target_amount += detail.team_target_monthly
+                        rec.target_amount = target_amount
+                    # 不选择业务员,直接取门店年度目标
+                    else:
+                        target_year = self.env['team.target.year'].search([('team_id', '=', rec.kunnr.id),
+                                                                           ('target_year', '=', rec.target_year)])
+                        target_amount = 0
+                        if target_year:
+                            for target in target_year:
+                                target_amount += target.invoiced_target_year
+                        rec.target_amount = target_amount
+            else:
+                pass
 
     @api.depends('target_year', 'target_month', 'kunnr', 'ywy')
     def _compute_jiesuan_amount(self):
         for rec in self:
-            if rec.kunnr:
-                kunnr_sql = "and kunnr = %s" % rec.kunnr.id
+            # 确认是查询目标完成占比报表
+            if rec.target_year:
+                if rec.kunnr:
+                    kunnr_sql = "and kunnr = %s" % rec.kunnr.id
+                else:
+                    kunnr_sql = ''
+                if rec.ywy:
+                    ywy_sql = "and ywy = '%s'" % rec.ywy.id
+                else:
+                    ywy_sql = ""
+                if rec.target_month:
+                    month = '%02d' % int(rec.target_month)
+                    time_sql = "TO_CHAR(\"LFADT\", 'YYYYMM') = '%s'" % (str(rec.target_year) + str(month))
+                    sql_str = "select * from outbound_final where %s %s %s" % (time_sql, kunnr_sql, ywy_sql)
+                else:
+                    time_sql = "TO_CHAR(\"LFADT\", 'YYYY') = '%s'" % (str(rec.target_year))
+                    sql_str = "select * from outbound_final where %s %s %s" % (time_sql, kunnr_sql, ywy_sql)
+                rec._cr.execute(sql_str)
+                res_amount = rec._cr.dictfetchall()
+                jiesuan_amount = 0
+                if res_amount:
+                    for res in res_amount:
+                        jiesuan_amount += res['jiesuanjine']
+                rec.jiesuan_amount = jiesuan_amount
             else:
-                kunnr_sql = ''
-            if rec.ywy:
-                ywy_sql = "and ywy = '%s'" % rec.ywy.id
-            else:
-                ywy_sql = ""
-            if rec.target_month:
-                month = '%02d' % int(rec.target_month)
-                time_sql = "TO_CHAR(\"LFADT\", 'YYYYMM') = '%s'" % (str(rec.target_year) + str(month))
-                sql_str = "select * from outbound_final where %s %s %s" % (time_sql, kunnr_sql, ywy_sql)
-            else:
-                time_sql = "TO_CHAR(\"LFADT\", 'YYYY') = '%s'" % (str(rec.target_year))
-                sql_str = "select * from outbound_final where %s %s %s" % (time_sql, kunnr_sql, ywy_sql)
-            rec._cr.execute(sql_str)
-            res_amount = rec._cr.dictfetchall()
-            jiesuan_amount = 0
-            if res_amount:
-                for res in res_amount:
-                    jiesuan_amount += res['jiesuanjine']
-            rec.jiesuan_amount = jiesuan_amount
+                pass
 
     # 获取查询视图的view_id,在js中访问该方法获取指定该视图id
     @api.model
